config/utils/scripts/reload_polybar: always do everything was wrong :melting_face:

+++ config/utils/scripts/reload-polybar.sh
@@ -1,7 +1,10 @@
 #!/usr/bin/env bash
 
-pkill -USR1 polybar
-for m in $(xrandr --query | grep " connected" | cut -d" " -f1); do
-    MONITOR=$m polybar --reload bar&
-done
+if pgrep "polybar" > /dev/null; then
+    pkill -USR1 polybar
+else
+    for m in $(xrandr --query | grep " connected" | cut -d" " -f1); do
+        MONITOR=$m polybar --reload bar&
+    done
+fi
